+++ src/config/configSchema.ts
@@ -83,6 +83,9 @@ export type ConfigurationSchema = {
   cloudUploadClips: boolean;
   language: string;
   hideEmptyCategories: boolean;
+  hardwareAcceleration: boolean;
+  recordCurrentRaidEncountersOnly: boolean;
+  uploadCurrentRaidEncountersOnly: boolean;
 };
 
 export type ConfigurationSchemaKey = keyof ConfigurationSchema;
@@ -508,6 +511,21 @@ export const configSchema = {
   },
   hideEmptyCategories: {
     description: Phrase.HideEmptyCategoriesDescription,
+    type: 'boolean',
+    default: false,
+  },
+  hardwareAcceleration: {
+    description: Phrase.HardwareAccelerationDescription,
+    type: 'boolean',
+    default: true,
+  },
+  recordCurrentRaidEncountersOnly: {
+    description: Phrase.RecordCurrentRaidsOnlyDescription,
+    type: 'boolean',
+    default: false,
+  },
+  uploadCurrentRaidEncountersOnly: {
+    description: Phrase.UploadCurrentRaidsOnlyDescription,
     type: 'boolean',
     default: false,
   },