+++ src/localisation/german.ts
@@ -105,7 +105,7 @@ const GERMAN: Translations = {
   [Phrase.SeparateBufferFolderLabel]: 'Separater Buffer Ordner',
   [Phrase.BufferFolderLabel]: 'Buffer Ordner',
   [Phrase.MaxDiskStorageLabel]: 'Maximale Speichergröße(GB)',
-  [Phrase.AppSettingsLabel]: 'Anwendung Einstellungen',
+  [Phrase.WindowsSettingsLabel]: 'Anwendung Einstellungen',
   [Phrase.RunOnStartupLabel]: 'Autostart',
   [Phrase.StartMinimizedLabel]: 'Minimiert starten',
   [Phrase.MinimizeOnQuitLabel]: 'Minimieren beim schließen',
@@ -415,6 +415,12 @@ const GERMAN: Translations = {
   [Phrase.ProcessVolumeDescription]: "Die Lautstärke der Anwendung in der Aufnahme, von 0 bis 1.",
   [Phrase.HideEmptyCategoriesLabel]: "Leere Kategorien ausblenden",
   [Phrase.HideEmptyCategoriesDescription]: "Blendet Kategorien im Seitenmenü aus, die keine Videos enthalten.",
+  [Phrase.HardwareAccelerationLabel]: "Hardwarebeschleunigung",
+  [Phrase.HardwareAccelerationDescription]: "Aktivieren Sie die hardwarebeschleunigte Darstellung der Anwendung. Dies wird den meisten Benutzern empfohlen, insbesondere denen mit AV1-Kodierung, kann aber auf manchen Systemen zu Problemen führen.",
+  [Phrase.RecordCurrentRaidsOnlyLabel]: "Nur aktuelle Raids aufzeichnen",
+  [Phrase.RecordCurrentRaidsOnlyDescription]: "Zeichne nur Raid-Begegnungen der aktuellen Stufe auf, dies gilt nur für Retail-Raid-Begegnungen.",
+  [Phrase.UploadCurrentRaidsOnlyLabel]: "Nur aktuelle Raids hochladen",
+  [Phrase.UploadCurrentRaidsOnlyDescription]: "Laden Sie nur Raid-Begegnungen der aktuellen Stufe hoch. Dies gilt nur für Einzelhandels-Raid-Begegnungen.",
 };
 
 export default GERMAN;