+++ src/localisation/korean.ts
@@ -105,7 +105,7 @@ const KOREAN: Translations = {
   [Phrase.SeparateBufferFolderLabel]: '별도의 임시 폴더',
   [Phrase.BufferFolderLabel]: '임시 폴더',
   [Phrase.MaxDiskStorageLabel]: '디스크 저장 용량 제한 (GB)',
-  [Phrase.AppSettingsLabel]: '윈도우 설정',
+  [Phrase.WindowsSettingsLabel]: '윈도우 설정',
   [Phrase.RunOnStartupLabel]: '시작시 실행',
   [Phrase.StartMinimizedLabel]: '트레이 시작',
   [Phrase.MinimizeOnQuitLabel]: '종료시 트레이',
@@ -414,7 +414,13 @@ const KOREAN: Translations = {
   [Phrase.AudioProcessDevicesDescription]: "선택한 스피커와 마이크 외에 오디오를 포함할 응용 프로그램",
   [Phrase.ProcessVolumeDescription]: "기록에서 응용 프로그램의 볼륨은 0에서 1까지입니다.",
   [Phrase.HideEmptyCategoriesLabel]: "빈 카테고리 숨기기",
-  [Phrase.HideEmptyCategoriesDescription]: "비디오가 없는 사이드 메뉴의 카테고리를 숨깁니다.",
+  [Phrase.HideEmptyCategoriesDescription]: "비디오가 없는 카테고리를 숨깁니다.",
+  [Phrase.HardwareAccelerationLabel]: "하드웨어 가속",
+  [Phrase.HardwareAccelerationDescription]: "프로그램의 하드웨어 가속 렌더링을 활성화합니다. 이 기능은 대부분의 사용자에게 권장되며, 특히 AV1 인코딩을 사용하는 경우에 유용하지만 일부 시스템에서는 문제가 발생할 수 있습니다. 이 변경 사항은 프로그램을 재시작해야 적용됩니다.",
+  [Phrase.RecordCurrentRaidsOnlyLabel]: "현재 시즌 레이드만 기록",
+  [Phrase.RecordCurrentRaidsOnlyDescription]: "현재 시즌의 레이드 전투만 기록합니다. 이는 본섭 레이드 전투에만 적용됩니다.",
+  [Phrase.UploadCurrentRaidsOnlyLabel]: "현재 시즌 레이드만 업로드",
+  [Phrase.UploadCurrentRaidsOnlyDescription]: "현재 시즌의 레이드 전투만 업로드합니다. 이는 본섭 레이드 전투에만 적용됩니다.",
 };
 
 export default KOREAN;